requirements.txt
# VideoShortsAgent - Python Agent 依赖
faster-whisper>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
